Improve EM850i support.
Per user manual ecoMAX 850i circuit 'support' parameter is binary only for H1. For H2 and H3 it can be one of the following: "Off", "On (radiators)", "On (floor)".

To facilitate this, filter mixer descriptions by index to make sure that 'Enable circuit' switch is added for the first circuit and 'Work mode' selector is added for second and third.

Related to: #42